% Auto-generated: do not edit by hand
\name{dashArcherContainer}

\alias{dashArcherContainer}

\title{DashArcherContainer component}

\description{
DashArcherContainer component is a wrapper for ArcherContainer component from react-archer module. See usage: https://github.com/pierpo/react-archer
}

\usage{
dashArcherContainer(children=NULL, arrowLength=NULL, arrowThickness=NULL, id=NULL, noCurves=NULL,
offset=NULL, strokeColor=NULL, strokeDasharray=NULL, strokeWidth=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{arrowLength}{Numeric. A size in px.}

\item{arrowThickness}{Numeric. A size in px.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{noCurves}{Logical. Set this to true if you want angles instead of curves.}

\item{offset}{Numeric. Optional number for space between element and start/end of stroke.}

\item{strokeColor}{Character. A color string '#ff0000'.}

\item{strokeDasharray}{Character. Adds dashes to the stroke. It has to be a string representing an array
of sizes. See https://www.w3schools.com/graphics/svg_stroking.asp}

\item{strokeWidth}{Numeric. A size in px.}
}
